requirements.txt
fastapi==0.116.1
httpx==0.28.1
Jinja2==3.1.4
PyYAML==6.0.2
pytest==8.4.1
python-multipart==0.0.9
ruff==0.12.7
